requirements.txt
pandas>=1.3.0
numpy>=1.20.0
Pillow>=8.0.0 
